Fix crossfader sweep, strip placement, ffmpeg fail-fast (#31)
* Fail fast with a clear error when ffmpeg is missing

probe archive, probe refit, and render all shell out to ffmpeg but only
discovered that after loading a program on the rig, wasting device time
before the FileNotFoundError surfaced. require_ffmpeg() now checks before
any device work starts.


* Sweep the crossfader instead of pinning it open

Pinning P12 at max for a whole program's archive pass produced false
blank-take failures (Bitcullis, Corollas): the picture depends on it like
any other effect. It now sweeps 0..100% via the same Sobol plan as every
other continuous parameter, forced continuous even on programs that report
it as unused ("Null 12").


* Move the gray-code strip to the bottom edge

The state-index and render timecode strips sat on the top edge of every
frame, where they can interfere with videomancer programs. Both moved to
the bottom edge: patterns.with_state_index/read_state_index/crop_strip,
the codeframes loop's body/strip layout, and render.py's independent
burn_timecode/read_timecode/crop_strip.


* Widen a uint8 read before pytest.approx to avoid NEP 50 underflow

decoded[16, i*16+8] is uint8; subtracting a Python int under NEP 50 stays
uint8 and can wrap (30 - 31 -> 255), failing a comparison that is actually
within tolerance. Cast to int first.


---------

+++ tests/probe/test_archive.py
@@ -473,7 +473,8 @@ def test_a_real_recording_reads_back_frame_for_frame_in_colour(tmp_path):
         decoded = reader.frame(index)
         assert decoded.dtype == np.uint8 and decoded.shape == (32, 64, CHANNELS)
         for i, colour in enumerate(BLOCKS[index:] + BLOCKS[:index]):
-            assert decoded[16, i * 16 + 8] == pytest.approx(colour, abs=4)
+            # uint8 minus a Python int can underflow (NEP 50); widen before comparing.
+            assert decoded[16, i * 16 + 8].astype(int) == pytest.approx(colour, abs=4)
     assert [f.shape for f in reader.stream()] == [(32, 64, CHANNELS)] * 4
     assert [f.shape for f in reader.stream(width=32)] == [reader.shape(32)] * 4
